+++ General/settings.lua
@@ -171,7 +171,7 @@ local function SettingsMenu_OnShow()
 		settingsFrame.title = settingsFrame:CreateFontString(nil, "OVERLAY");
 		settingsFrame.title:SetFontObject("GameFontHighlight");
 		settingsFrame.title:SetPoint("CENTER", settingsFrame.TitleBg, "CENTER", 5, 0);
-		settingsFrame.title:SetText(L["ADDON_NAME_SETTINGS"]);
+		settingsFrame.title:SetText(L["ADDON_NAME_SETTINGS"] .. " - " .. L["RELEASE"]);
 	end
 	
 	----- START TAB1 -----
@@ -191,18 +191,18 @@ local function SettingsMenu_OnShow()
 		LastSeenTbl.tab1.rankLabel:SetText(L["RANK"] .. LastSeenTbl.GetPlayerRank());
 	end
 	
-	if not LastSeenTbl.tab1.versionLabel then
+	--[[if not LastSeenTbl.tab1.versionLabel then
 		LastSeenTbl.tab1.versionLabel = LastSeenTbl.tab1:CreateFontString(nil, "OVERLAY");
 		LastSeenTbl.tab1.versionLabel:SetFontObject("GameFontHighlight");
 		LastSeenTbl.tab1.versionLabel:SetPoint("TOPRIGHT", LastSeenTbl.tab1.itemsSeenLabel, 265, 0);
 		LastSeenTbl.tab1.versionLabel:SetFont("Fonts\\Arial.ttf", 8);
 		LastSeenTbl.tab1.versionLabel:SetText(L["RELEASE"]);
-	end
+	end]]--
 
 	if not LastSeenTbl.tab1.releaseDateLabel then
 		LastSeenTbl.tab1.releaseDateLabel = LastSeenTbl.tab1:CreateFontString(nil, "OVERLAY");
 		LastSeenTbl.tab1.releaseDateLabel:SetFontObject("GameFontHighlight");
-		LastSeenTbl.tab1.releaseDateLabel:SetPoint("TOP", LastSeenTbl.tab1.versionLabel, "BOTTOM", -15, -5);
+		LastSeenTbl.tab1.releaseDateLabel:SetPoint("TOPRIGHT", LastSeenTbl.tab1.itemsSeenLabel, 265, 0);
 		LastSeenTbl.tab1.releaseDateLabel:SetFont("Fonts\\Arial.ttf", 8);
 		LastSeenTbl.tab1.releaseDateLabel:SetText("24.09.2019");
 	end